delay added for JC write

+++ Core/Inc/joint_config.h
@@ -10,11 +10,11 @@
 
 struct joint_config
 {
-	float number;
-	float motor_type;
-	float upper_limit;
-	float lower_limit;
-	float zero;
+	int32_t number;
+	int32_t motor_type;
+	int32_t upper_limit;
+	int32_t lower_limit;
+	int32_t zero;
 };
 
 typedef struct joint_config joint_config;
